some further warning cleanup

+++ src/Igor/src/iWidgetGraph.h
@@ -276,7 +276,7 @@ namespace Igor
 
         /*! button height definition
         */
-        const float32 _buttonHeight = 20;
+        const int32 _buttonHeight = 20;
 
         /*! makes some calculations before we can draw
         */
